Update
1. Fix cast warning
2. iopmp: Use entry_offset
3.Remove reduandent define

+++ src/utility/plic.h
@@ -10,12 +10,12 @@
 #define PLIC_CLAIM_COMPLETE (PLIC_BASE + 0x200004) // For hart 0
 
 __attribute__((always_inline)) static inline void plic_enable_interrupt(uint32_t irq, uint32_t priority) {
-    *((volatile uint32_t *)(PLIC_PRIORITY_BASE + irq * 4)) = priority;
+    *((volatile uint32_t *)(PLIC_PRIORITY_BASE + (long)irq * 4)) = priority;
 
     // Enable the interrupt for hart 0
     uint32_t reg_offset = irq / 32;
     uint32_t bit_offset = irq % 32;
-    volatile uint32_t *enable_reg = (volatile uint32_t *)(PLIC_ENABLE_BASE + reg_offset * 4);
+    volatile uint32_t *enable_reg = (volatile uint32_t *)(long)(PLIC_ENABLE_BASE + (long)reg_offset * 4);
     *enable_reg |= (1 << bit_offset);
 }
 
@@ -23,7 +23,7 @@ __attribute__((always_inline)) static inline void plic_disable_interrupt(uint32_
     // Disable the interrupt for hart 0
     uint32_t reg_offset = irq / 32;
     uint32_t bit_offset = irq % 32;
-    volatile uint32_t *enable_reg = (volatile uint32_t *)(PLIC_ENABLE_BASE + reg_offset * 4);
+    volatile uint32_t *enable_reg = (volatile uint32_t *)(PLIC_ENABLE_BASE + (long)reg_offset * 4);
     *enable_reg &= ~(1 << bit_offset);
 }
